pyproject.toml
[project]
name = "plainconf"
version = "0.1.11"
description = "Minimal configuration handling supporting hashicorp vault."
authors = [{name="Caspar David Dzikus", email="caspar@kodekitchen.com"}]
classifiers = ["Development Status :: 2 - Pre-Alpha"]
readme = "README.md"
dependencies = ["python-dotenv", "hvac", "tomli_w", "crpytography"]
keywords = ["configuration", "vault"]

[project.urls]
repository = "https://github.com/kodekitchen/plainconf"
bugtracker = "https://github.com/kodekitchen/plainconf/issues"
